backend/otif-model/requirements.txt
pandas
numpy
lightgbm
scikit-learn
shap
sqlalchemy
pyodbc
joblib
streamlit
pyyaml
python-pptx
matplotlib
seaborn
plotly
fpdf2
openai
python-dotenv
fastapi 
uvicorn[standard] 
sqlalchemy 
passlib[bcrypt] 
python-jose 
pydantic
pydantic-settings
pydantic[email]
python-multipart